requirements.txt
nose
expecter
pyzmq
fuzzywuzzy
requests
simplejson
